packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qagentic-pytest"
version = "0.1.0"
description = "AI-Powered Test Intelligence SDK for Python - pytest plugin with automatic failure analysis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "QAagentic Team", email = "team@qagentic.io" }
]
keywords = [
    "testing",
    "pytest",
    "test-reporting",
    "ai",
    "test-analytics",
    "quality-assurance",
    "automation",
    "allure",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/qagentic/qagentic-sdk"
Documentation = "https://docs.qagentic.io"
Repository = "https://github.com/qagentic/qagentic-sdk"
Issues = "https://github.com/qagentic/qagentic-sdk/issues"

[project.entry-points.pytest11]
qagentic = "qagentic.pytest_plugin"

[tool.hatch.build.targets.wheel]
packages = ["src/qagentic"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
